+++ Module-API/src/main/java/depromeet/api/domain/feed/usecase/GetChallengeFeedUseCase.java
@@ -4,11 +4,11 @@ package depromeet.api.domain.feed.usecase;
 import depromeet.api.domain.feed.dto.response.GetChallengeFeedResponse;
 import depromeet.api.domain.feed.mapper.RecordListMapper;
 import depromeet.common.annotation.UseCase;
-import depromeet.domain.challenge.adaptor.UserChallengeAdaptor;
 import depromeet.domain.record.adaptor.RecordAdaptor;
 import depromeet.domain.record.domain.Record;
 import depromeet.domain.user.adaptor.UserAdaptor;
 import depromeet.domain.user.domain.User;
+import depromeet.domain.userchallenge.adaptor.UserChallengeAdaptor;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
