+++ Module-API/src/main/java/depromeet/api/domain/feed/usecase/GetChallengeProceedingInfoUseCase.java
@@ -4,10 +4,10 @@ package depromeet.api.domain.feed.usecase;
 import depromeet.api.domain.feed.dto.response.GetChallengeProceedingInfoResponse;
 import depromeet.api.domain.feed.mapper.UserChallengeMapper;
 import depromeet.common.annotation.UseCase;
-import depromeet.domain.challenge.adaptor.UserChallengeAdaptor;
-import depromeet.domain.challenge.domain.UserChallenge;
 import depromeet.domain.user.adaptor.UserAdaptor;
 import depromeet.domain.user.domain.User;
+import depromeet.domain.userchallenge.adaptor.UserChallengeAdaptor;
+import depromeet.domain.userchallenge.domain.UserChallenge;
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
