+++ Module-Common/src/main/java/depromeet/common/exception/CustomExceptionStatus.java
@@ -72,9 +72,8 @@ public enum CustomExceptionStatus {
     DUPLICATED_EMOJI(false, 2301, "이미 반응했던 이모지가 중복됩니다."),
 
     // userChallenge
-    UNPARTICIPATED_CHALLENGE_USER(false, 2400, "챌린지에 참여하지 않는 사용자입니다.");
-    USER_NOT_PARTICIPATED_IN_CHALLENGE(false, 2400, "해당 챌린지에 참여중인 유저가 아닙니다.");
-    PROGRESS_INFO_NOT_FOUND(false, 2401, "챌린지가 존재하지 않거나, 해당 챌린지에 참여중이 아닙니다."),
+    USER_NOT_PARTICIPATED_IN_CHALLENGE(false, 2400, "해당 챌린지에 참여중인 유저가 아닙니다."),
+    PROGRESS_INFO_NOT_FOUND(false, 2401, "챌린지가 존재하지 않거나, 해당 챌린지에 참여중이 아닙니다.");
 
     private final boolean isSuccess;
     private final int code;
