+++ Module-Domain/src/main/java/depromeet/domain/challenge/domain/Challenge.java
@@ -8,6 +8,7 @@ import depromeet.domain.challenge.exception.ChallengeCannotBeUpdatedAfterStartEx
 import depromeet.domain.config.BaseTime;
 import depromeet.domain.keyword.domain.Keyword;
 import depromeet.domain.rule.domain.ChallengeRule;
+import depromeet.domain.userchallenge.domain.UserChallenge;
 import java.time.LocalDate;
 import java.util.List;
 import javax.persistence.*;
