--- Module-Domain/src/main/java/depromeet/domain/challenge/repository/UserChallengeRepository.java
@@ -1,18 +1,0 @@
-package depromeet.domain.challenge.repository;
-
-
-import depromeet.domain.challenge.domain.UserChallenge;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
-public interface UserChallengeRepository
-        extends JpaRepository<UserChallenge, Long>, UserChallengeCustomRepository {
-
-    @Query(
-            value =
-                    "SELECT uc FROM UserChallenge uc WHERE uc.user.id = :uid AND uc.challenge.id = :cid")
-    Optional<UserChallenge> findByUserIdAndChallengeRoomId(
-            @Param("uid") Long userId, @Param("cid") Long challengeRoomId);
-}
