--- Module-Domain/src/main/java/depromeet/domain/userchallenge/exception/UnparticipatedChallengeUserException.java
@@ -1,13 +1,0 @@
-package depromeet.domain.userchallenge.exception;
-
-
-import depromeet.common.exception.CustomException;
-import depromeet.common.exception.CustomExceptionStatus;
-
-public class UnparticipatedChallengeUserException extends CustomException {
-    public static final CustomException EXCEPTION = new UnparticipatedChallengeUserException();
-
-    private UnparticipatedChallengeUserException() {
-        super(CustomExceptionStatus.UNPARTICIPATED_CHALLENGE_USER);
-    }
-}
